bot/requirements.txt
python-dotenv>=0.9.9,<1.0.0
faiss-cpu>=1.11.0,<2.0.0
numpy>=1.23.0,<2.0.0
torch>=2.6.0,<3.0.0
transformers>=4.49.0,<5.0.0

langchain-core>=0.3.25,<1.0.0
langchain-community>=0.3.24,<1.0.0
langchain-mistralai>=0.2.10,<1.0.0

tritonclient[http]>=2.27.0,<3.0.0

python-telegram-bot>=22.1.0,<23.0.0
